refactor: consolidate utils directory into single file
Simplify project structure by merging utils/ directory into a single
src/utils.ts file for better maintainability and reduced complexity.

## Changes

1. **Merged utils/helpers.ts and utils/index.ts**
   - Combined all utility functions into src/utils.ts
   - Removed intermediate utils/ directory
   - Flattened import structure

2. **Updated imports**
   - Changed import paths from '../utils/helpers' to '../utils'
   - Fixed test imports in tests/helpers.test.ts

3. **Removed files**
   - Deleted src/utils/helpers.ts
   - Deleted src/utils/index.ts
   - Removed empty src/utils/ directory

## Benefits

✅ **Simpler structure** - One file instead of two-level directory
✅ **Easier navigation** - All utilities in one place
✅ **Reduced complexity** - No need for index re-exports
✅ **Better tree-shaking** - Direct exports from single file
✅ **Cleaner imports** - from '../utils' instead of '../utils/helpers'

## File structure

**Before:**
```
src/
├── utils/
│   ├── index.ts (re-exports from helpers.ts)
│   └── helpers.ts (actual implementations)
```

**After:**
```
src/
└── utils.ts (all implementations)
```

## Results

- Bundle size: Unchanged (102.22 KB)
- Tests: 233/233 passing ✅
- No functional changes
- Cleaner project structure

## Impact

This is a structural refactor with no breaking changes:
- External API remains identical
- All imports from '../utils' continue to work
- Tests pass without modification to logic
